Notion and user config (#7)
* added: notion, yaml user config, words filtering

* fixes: config path problem

* added: backoff on requests

+++ src/handlers/notes.py
@@ -14,6 +14,9 @@ class MessageServiceProtocol(typing.Protocol):
 
 
 class NotesServiceProtocol(typing.Protocol):
+    delete_done_notes: bool = False
+    start_words: list[str] = []
+
     async def create_note(self, text: str) -> None:
         ...
 
@@ -27,17 +30,26 @@ class NotesServiceProtocol(typing.Protocol):
         ...
 
 
+class NotesFilterProtocol(typing.Protocol):
+    def get_needed_to_create_notes(self, text: str) -> list[NotesServiceProtocol]:
+        ...
+
+
 class NotesHandler:
-    def __init__(self, message_service: MessageServiceProtocol) -> None:
+    def __init__(self, message_service: MessageServiceProtocol, filter_class: type[NotesFilterProtocol]) -> None:
         self._message_service = message_service
+        self._filter_class = filter_class
         self._notes_services: list[NotesServiceProtocol] = []
 
-    def with_notes_service(self, notes_service: NotesServiceProtocol) -> typing.Self:
+    def with_notes_service(self, notes_service: NotesServiceProtocol,
+                           delete_done_notes: bool, start_words: list[str]) -> typing.Self:
+        notes_service.delete_done_notes = delete_done_notes
+        notes_service.start_words = start_words
         self._notes_services += [notes_service]
         return self
 
     async def _create_notes(self, text: str) -> None:
-        for notes_service in self._notes_services:
+        for notes_service in self._filter_class(self._notes_services).get_needed_to_create_notes(text):
             await notes_service.create_note(text)
 
     async def _get_notes(self) -> str:
@@ -59,6 +71,7 @@ class NotesHandler:
     async def delete_done_notes(self) -> None:
         logger.debug('Delete done notes')
         for notes_service in self._notes_services:
-            note_ids = await notes_service.get_done_note_ids()
-            for note_id in note_ids:
-                await notes_service.delete_note(note_id)
+            if notes_service.delete_done_notes:
+                note_ids = await notes_service.get_done_note_ids()
+                for note_id in note_ids:
+                    await notes_service.delete_note(note_id)
